chore: render proper icons for due dates
